pyproject.toml
[project]
name = ""
version = "0.1.0"
description = ""
authors = [{ name = "Repay Team", email = "repay@repay.com" }]
requires-python = ">=3.12,<3.13"
readme = "README.md"

dependencies = [
    "sqlalchemy>=2.0.36,<3",
    "fastapi>=0.115.6,<0.116",
    "pydantic-settings>=2.7.0,<3",
    "pyyaml>=6.0.2,<7",
    "uvicorn>=0.34.0,<0.35",
    "alembic>=1.14.0,<2",
    "asyncpg>=0.30.0,<0.31",
    "greenlet>=3.1.1,<4",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.4,<9",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.8.4,<0.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
package = false

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"


[tool.ruff]
src = ["src"]
lint.select = [
    "FAST", "ASYNC", "ANN", "S", "B", "A", "COM", "C4",
    # "DTZ",
    "T10", "EM", "EXE", "ISC", "ICN", "INP", "PIE", "T20", "PYI", "Q", "RSE", "RET", "SLF", "SIM", "SLOT", "TID",
    "TC", "ARG", "PTH", "FLY", "I", "NPY", "PD", "N", "PERF", "E", "W", "F", "PLC", "PLE", "PLR", "PLW", "UP", "FURB",
    "LOG", # ?
    "RUF",
]
lint.ignore = ["RUF100", "PLR0913", "ANN401", "S608", "S324"]
lint.fixable = ["ALL"]
lint.unfixable = []
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
lint.allowed-confusables = ["а", "А", "б", "Б", "в", "В", "г", "Г", "е", "Е", "з", "З", "и", "И", "к", "К", "м", "М", "н", "Н", "о", "О", "р", "Р", "с", "С", "у", "У", "ф", "Ф", "х", "Х"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "src/infra/postgres/alembic",
]
line-length = 120
target-version = "py312"

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["src"]
